Improved Responsiveness and login auth
improved responsiveness for navbar. made login/logout responsive to when authenticated (i think). it's far from perfect, but it's much more responsive now. i couldn't get it to change to a hamburger at 1250px, which is when it seems to have a tough time. probably needs to be tweaked once again, but i'd rather give that to someone a bit better at frontend